Added some codes to SearchViewModel class and to list_item_search_query_summary.xml to determine if the Contact View summary should be visible or not in the display layout based on the total number of Contact
Previous changes:
Added some codes to SearchViewModel class and to list_item_search_query_summary.xml to determine if the Company View summary should be visible or not in the display layout based on the total number of Company
Added some codes to SearchViewModel class and to list_item_search_query_summary.xml to determine if the Project View summary should be visible or not in the display layout based on the total number of Project.
Modified the code and xml layout to match the display regarding about the Contacts section based on the design from the sketch file.
Modified the search bar layout section to match the design based on the sketch file.