Created the SearchFilterMPFOwnerTypeViewModel class for Owner Type section. Added codes for selecting a checkbox item. Modified the corresponding xml layout for the Owner type and applied data binding for the model and view layout.
Created the SearchFilterMPFWorkTypeViewModel class for Work Type section. Added codes for selecting a checkbox item.
Modified the corresponding xml layout for the Work type and applied data binding for the model and view layout.

Created the SearchFilterMPFJurisdictionViewModel class for Jurisdiction section. Added codes for selecting a checkbox item.
Modified the corresponding xml layout for the Jurisdiction and applied data binding for the model and view layout.

Created the SearchFilterMPFStageViewModel class for Stage section. Added codes for selecting a checkbox item.
Modified the corresponding xml layout for the Stage and applied data binding for the model and view layout.

Modified the onActivityResult() method of the SearchFilterMPSActivity class to include the data-binding processing of the Owner Type Section, Work Type section, Jurisdiction section and the Stage section.

Last changes:
1. SearchFilterMPFLocationViewModel - for Project Location
2. SearchFilterMPFTypeViewModel class - for Project Type
3. SearchFilterMPFValueViewModel class - for Project Value
4. SearchFilterMPFUpdatedWithinViewModel class - for Project Updated Within
5. SearchFilterMPFBiddingWithinViewModel class - for Project BiddingWithin
6. SearchFilterMPFBHViewModel class - for Project B/H
7. SearchFilterMPFWorkTypeActivity class - for Project Work Type activity class
8. activity_search_filter_mpfwork-type.xml - layout for Project Work Type
9. Modified all related layout  for the newly created view model classes for project location, type, value, updated within, bidding within and b/h.
10. Modified the activity_search_filter_mps30.xml to databind the user's selected items for project location, type, value, updated within, bidding within and b/h.
11. Modified the SearchFilterMPSActivity class to add codes inside the onActivityResult  for project location, type, value, updated within, bidding within and b/h.