Changed some codes and added comment in JurisdictionAdapter class and SearchFilterMPFJurisdictionActivity2 class
 Previous changes:
Created and implemented the SearchMPFJurisdictionActivity2 class for implementing the recycle view of new layout for jurisdiction.
Created the activity_search_filter_mpfjurisdiction2 xml layout for SearchMPFJurisdictionActivity2 class.
Created and implemented the JurisdictionAdapter class for recycle view adapter for jurisdiction.
Created the list_item_mpfjurisdiction_parent xml layout for the parent items of jurisdiction.
Created the list_item_mpfjurisdiction_local xml layout for the local items of jurisdiction.
Created the list_item_mpfjurisdiction_district xml layout for the district council items of jurisdiction.
Modified the SearchSaved POJO model class to replace the SearchFilter type of one variable to Filter type variable.
Created and implemented the Filter class for the SearchSaved class.
Changed the content of the SearchFilter class based on the correct output coming from the json data.
Added POJOs to SearchViewModel class.

Reduced the size of down arrow image closer to the sketch design in styles.xml.
Modified the SearchDomain class to monitor the content of the main class of the jurisdiction, project types & stages.
Modified the SearchViewModel class to change the local main object variable of jurisdiction  , project types & stages into public class main variable so it would be correctly processed of getting the result data from the web only once.
Modified the activity_search_filter_mp30.xml to include this added style for filter value section and removed the excess horizontal line.
Added style for Search filter field value section in style.xml
Changed the correct color in the rectangular border of the search filter field value.