Fixed bug case 6054 - search returns inconsistent result.
In SearchDomain class:
Created three static Call type variables for Project, Company and Contacts. This will use to monitor if there's a processing of previous HTTP request.
Modified the getSearchProjectQuery(), getSearchCompanyQuery() and getSearchContactQuery() to replace the local Call type variable with class Call type variable.

In SearchViewModel class:
Modified the updateViewQuery() to add the cancelling of  processing the HTTP request if the Call object exists for Project, Company and Contact.
Re-positioned the calling of the checkTotal() to correctly evaluate the total list of Project, Company and Contact.