Fix completed for UAT 8168 - Displaying next batch results for Project, Company and Contact.
In SearchViewModel class, added new boolean variables to monitor the loading of next batch results for contact. added variables to monitor the skip count for loading the next batch results for contact. Created getContactAdditionalData() to process the loading of the next batch results for contact. Modified the initializeAdapterContactQueryAll() to include the code for implementing the onScroll event and for monitoring if the next contact batch results is needed to add to recycle view list item.
In SearchItemRecentViewModel class, modified the constructor to include the SearchViewModel type parameter for Contact. Modified the onContactClick() to include the condition for setting the detailVisible to true to be used so that the app knows that the current display of the view is outside the search section view.
In SearchSummaryContactRecyclerViewAdapter class and SearchAllContactRecyclerViewAdapter class, Added a SearchActivity type variable to hold the SearchActivity object to be used for passing the SearchViewModel object to the constructor of the SearchItemRecentViewModel class.
In SearchRecentRecyclerViewAdapter class, added the passing of the SeachViewModel object to the last parameter of the constructor for the contact.